Simplify Windows proxy handling
Each supported proxy requires explicit support in UnityDoorstop, so why
are we bothering to half-support proxying to arbitrary DLLs? I've
simplified the code to explicitly check if our name is winhttp.dll and
if it is, to search the system directory for winhttp.dll, without
checking for an `_alt.dll` file.